requirements.txt
customtkinter
pdfplumber
pdf2image
Pillow
opencv-python
numpy
